Refactor NPC form and add edit workflow
